options/posix: disable pthreads linux extensions without the linux option
