+++ tests/meson.build
@@ -92,6 +92,7 @@ all_test_cases = [
 	'glibc/error_at_line',
 	'linux/xattr',
 	'linux/pthread_setname_np',
+	'linux/pthread_attr',
 ]
 
 if host_machine.system() == 'linux'
